+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/rule_evaluation_non_decomposable_complete.hpp
@@ -18,9 +18,9 @@ namespace boosting {
     class NonDecomposableCompleteRuleEvaluationFactory final : public INonDecomposableRuleEvaluationFactory {
         private:
 
-            const float64 l1RegularizationWeight_;
+            const float32 l1RegularizationWeight_;
 
-            const float64 l2RegularizationWeight_;
+            const float32 l2RegularizationWeight_;
 
             const Blas& blas_;
 
@@ -38,7 +38,7 @@ namespace boosting {
              * @param lapack                    An reference to an object of type `Lapack` that allows to execute BLAS
              *                                  routines
              */
-            NonDecomposableCompleteRuleEvaluationFactory(float64 l1RegularizationWeight, float64 l2RegularizationWeight,
+            NonDecomposableCompleteRuleEvaluationFactory(float32 l1RegularizationWeight, float32 l2RegularizationWeight,
                                                          const Blas& blas, const Lapack& lapack);
 
             std::unique_ptr<IRuleEvaluation<DenseNonDecomposableStatisticVector<float64>>> create(
